refactor: BookedTicket 생성 시점 수정
- BookingComplete 시작 시점 -> BookingSeat 완료 시점으로 변경
