+++ app/src/main/java/woowacourse/movie/ui/complete/contract/BookingCompleteContract.kt
@@ -6,8 +6,6 @@ interface BookingCompleteContract {
     interface Presenter {
         fun loadBookedTicket(bookedTicket: BookedTicket)
 
-        fun insertBookedTicket()
-
         fun refreshBookedTicketDisplay()
 
         fun refreshTicketPrice()
