+++ app/src/main/java/woowacourse/movie/ui/seat/contract/BookingSeatContract.kt
@@ -1,16 +1,18 @@
 package woowacourse.movie.ui.seat.contract
 
 import woowacourse.movie.domain.model.movie.Headcount
+import woowacourse.movie.domain.model.theater.BookedTicket
 import woowacourse.movie.domain.model.theater.Seat
-import woowacourse.movie.domain.model.theater.Seats
 import woowacourse.movie.domain.model.theater.Theater
+import java.time.LocalDateTime
 
 interface BookingSeatContract {
     interface Presenter {
         fun loadState(
             theater: Theater,
             headcount: Headcount,
-            title: String
+            title: String,
+            bookedDateTime: LocalDateTime,
         )
 
         fun refreshTotalPrice()
@@ -20,6 +22,8 @@ interface BookingSeatContract {
         fun selectSeat(seat: Seat)
 
         fun refreshConfirmButton()
+
+        fun insertBookedTicket()
 
         fun completeBookingSeat()
     }
@@ -36,11 +40,6 @@ interface BookingSeatContract {
 
         fun setConfirmButton(isEnabled: Boolean)
 
-        fun startBookingCompleteActivity(
-            movieTitle: String,
-            headcount: Headcount,
-            seats: Seats,
-            theater: Theater,
-        )
+        fun startBookingCompleteActivity(bookedTicket: BookedTicket)
     }
 }
